tf cast to float32

+++ task4/custom_objects_2.py
@@ -129,7 +129,7 @@ class SiameseModel5(Model):
 
         y_true = tf.cast(y_true,tf.float32)
         abs_diff_sum = tf.math.reduce_sum(tf.math.abs(y_pred-y_true))
-        n_y_true = tf.size(tf.reshape(y_true,[-1]))
+        n_y_true = tf.cast(tf.size(tf.reshape(y_true,[-1])),tf.float32)
         acc = (n_y_true-abs_diff_sum)/n_y_true
         return acc
